Started writing some code for the daemon. Created a sequence diagram to demonstrate about how things will work.
Ran into some trouble when I looked at how the database will interact with the program to the hardware. I expect that each function to perform some hardware action will need to be written seperatly than the program, as we expected with the database. When we were designing the database however, we were planning on having a seperate program to deal with the hardware (which we call via a shell command). Although that option still exists, it creates a lot more points of failure and it seems very hackish. I am going to look into having each piece of hardware be managed by a dynamic library. The only trouble with that system would be that the dll's would need to interact with the hardware through the queues run by the daemon. This is a very similiar to concept to some code I wrote for a class a while back, I will dig up that code and add it to the repo.