Wrote some code, found an eclipse plugin to make Qt more useable.
The code was mostly for logging and what not, but should be all done. It is a big stepping stone, as most of the other threads will use a similiar architecture. The next most important bit is to get people to begin writing code; the standard header should look like...
void somefunc(void *data);